Move protractor from peerDependencies to devDependencies
It is devDependencies because it is needed to test the plugin itself.
The version for devDependencies is at least 0.14.0 because testConf.js
needs to change due to new webdriver-manager script in protractor v0.14.0

It is no longer peerDependencies because protractor is officially
suggested to be installed globally. Listed it in peerDependencies will
have npm installed it locally anyway and the global one will not get
used by this plugin. If user still want to install protractor locally,
list it as dependencies in user's package.json.